fix del post

+++ lib/view/component/post_newScreen.dart
@@ -201,15 +201,36 @@ class _PostScreenState extends State<PostScreenNew> {
                             );
                           }
                           else if (value == 'Xóa câu hỏi') {
-                            // Hành động khi chọn "Xóa câu hỏi"
-                            // hàm xóa
-                            await postController.deletePost(postHT!.id, context);
-
-                            //ửi biến về cho trang truoc đó load lại
-                            widget.onReportAction();
-
-
+                            // Hiển thị hộp thoại xác nhận
+                            showDialog(
+                              context: context,
+                              builder: (BuildContext context) {
+                                return AlertDialog(
+                                  title: Text("Xác nhận"),
+                                  content: Text("Bạn có chắc chắn muốn xóa câu hỏi này không?"),
+                                  actions: <Widget>[
+                                    TextButton(
+                                      child: Text("Hủy"),
+                                      onPressed: () {
+                                        Navigator.of(context).pop(); // Đóng hộp thoại và hủy hành động
+                                      },
+                                    ),
+                                    TextButton(
+                                      child: Text("Xóa"),
+                                      onPressed: () async {
+                                        Navigator.of(context).pop(); // Đóng hộp thoại
+                                        // Thực hiện hành động xóa
+                                        await postController.deletePost(postHT!.id, context);
+                                        // Gửi biến về cho trang trước đó load lại
+                                        widget.onReportAction();
+                                      },
+                                    ),
+                                  ],
+                                );
+                              },
+                            );
                           }
+
                         },
                         itemBuilder: (BuildContext context) {
                           return postHT!.createBy!.id! == loginController.idMe.value ?{'Truy cập Khoa', 'Xóa câu hỏi'}.map((String choice) {
